CEDS-2681 Refactor Fiscal Information page
